requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.1.0=pypi_0
bzip2=1.0.8=h4bc722e_7
ca-certificates=2024.12.14=hbcca054_0
certifi=2024.12.14=pypi_0
charset-normalizer=3.4.0=pypi_0
contextlib2=21.6.0=pypi_0
contourpy=1.1.1=pypi_0
cycler=0.12.1=pypi_0
cython=3.0.0=pypi_0
einops=0.8.0=pypi_0
fightingcv-attention=1.0.0=pypi_0
fonttools=4.55.3=pypi_0
h5py=3.11.0=pypi_0
idna=3.10=pypi_0
imageio=2.35.1=pypi_0
importlib-resources=6.4.5=pypi_0
kiwisolver=1.4.7=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.43=h712a8e2_2
libblas=3.9.0=26_linux64_openblas
libcblas=3.9.0=26_linux64_openblas
libffi=3.4.2=h7f98852_5
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgfortran=14.2.0=h69a702a_1
libgfortran5=14.2.0=hd5240d6_1
libgomp=14.2.0=h77fa898_1
liblapack=3.9.0=26_linux64_openblas
liblzma=5.6.3=hb9d3cd8_1
liblzma-devel=5.6.3=hb9d3cd8_1
libnsl=2.0.1=hd590300_0
libopenblas=0.3.28=pthreads_h94d23a6_1
libsqlite=3.47.2=hee588c1_0
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=14.2.0=h4852527_1
libuuid=2.38.1=h0b41bf4_0
libzlib=1.3.1=hb9d3cd8_2
matplotlib=3.7.5=pypi_0
ml-collections=0.1.1=pypi_0
ncurses=6.5=he02047a_1
networkx=3.1=pypi_0
numpy=1.24.4=py38h59b608b_0
opencv-python=4.8.0.74=pypi_0
openssl=3.4.0=hb9d3cd8_0
packaging=24.2=pypi_0
pillow=10.4.0=pypi_0
pip=24.3.1=pyh8b19718_0
pydensecrf=1.0rc3=py38h495e488_6
pyparsing=3.1.4=pypi_0
python=3.8.13=ha86cf86_0_cpython
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.8=5_cp38
pywavelets=1.4.1=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h8228510_1
requests=2.32.3=pypi_0
scikit-image=0.21.0=pypi_0
scipy=1.10.1=pypi_0
setuptools=75.3.0=pyhd8ed1ab_0
six=1.17.0=pypi_0
sqlite=3.47.2=h9eae976_0
tifffile=2023.7.10=pypi_0
tk=8.6.13=noxft_h4845f30_101
torch=1.12.1+cu113=pypi_0
torchaudio=0.12.1+cu113=pypi_0
torchvision=0.13.1+cu113=pypi_0
tqdm=4.67.1=pypi_0
typing-extensions=4.12.2=pypi_0
urllib3=2.2.3=pypi_0
wheel=0.45.1=pyhd8ed1ab_0
xz=5.6.3=hbcc6ac9_1
xz-gpl-tools=5.6.3=hbcc6ac9_1
xz-tools=5.6.3=hb9d3cd8_1
zipp=3.20.2=pypi_0
